backend update

+++ scaling_relations/xrayluminosity_halomass.py
@@ -131,7 +131,7 @@ def process_single_halo(
         data_T,
         data.gas.element_mass_fractions
     )
-    emissivities = unyt.unyt_array(10 ** emissivities, 'erg/s/cm**3')
+    emissivities = unyt.unyt_array((10 ** emissivities.astype(np.float32)).astype(np.float64), 'erg/s/cm**3')
     emissivities = emissivities.to('erg/s/Mpc**3')
 
     # Compute X-ray luminosities
